Move Text._thisFont creation into init
The Text component was creating a this._textFont object outside of init.
This was causing the structure to be shared with all instances of Text.
Whenever one instance changed the font it would be changed for all
instances.

This commit moves the initialization into init() so each instance will
get its own copy.