Adds Crafty.viewport.absoluteScale().
The existing scale() method applies the new scale as a multiplier
against the old scale. This adds an absoluteScale() method that sets
the scale absolutely, regardless of the old scale. Calling .scale(0.5)
twice results in a scale of 0.25. Calling .absoluteScale(0.5) twice
results in a scale of 0.5. This is useful when scaling the stage in
response to browser window resize events (see example code in comment
header).