Use floor/ceiling instead of round when calculating MBR.
To ensure that the rendered entity lies within the MBR, we need to take the floor of the minimum coords, and the ceiling of the maximum.  Consider: if minx = 0.7, and maxx=9.3, then with rounding the calculated width will be 8, which is not large enough to contain the true width of 8.6.