Add get() method to crafty core
This adds a get method that will convert selections to arrays of entities.  It can also be used to fetch a particular specified entity by index.  It will work as expected an a single entity.  This is modeled directly after jquery's get method: http://api.jquery.com/jQuery.get/

Documentation about the new feature is added (it is now referenced by the Crafty Core docs, as well as  .toArray).  Includes a few tests added to core.html.

This fixes #349