Switch redraw event name to "Invalidate"
Currently we use "Change" to indicate both when an attribute has been changed by `attr`, and to indicate when an entity needs to be redrawn.
This patch switches the name of the latter type of event to "Invalidate".

Since every other instance of "Invalidate" doesn't pass any date, this no longer passes the old coordinates when an entity is invalidated becuase of a move -- the event "Move" already covers that case.