Remove unwanted mobile behavior
Removing scroll on viewport is not something you want as it also happens on a iPad for an example.
The window.scrollTo does not work on iOS7 anymore, so that is also useless now, also because it will not scroll if the viewport height is equal to the document height.
Setting that the touchmove on is ignored on the Crafty element and not on the whole page, so that the webpage is still functional.