Fix DOM coordinates translation on Chrome 32
It happened that on the latest Chrome (32.0.1700.77) both
document.body.scrollTop and document.documentElement.scrollTop were
valorized, this caused a wrong computation of translated position of mouse
pointer with a missing trigger of a `Click` event on a DOM entity.

Note: I setup a separate test page for this test case to be used only with
*real browsers* as I found really difficult to handle page scrolling with
phantomjs.

More details about it at https://github.com/ariya/phantomjs/issues/10162
and https://github.com/gruntjs/grunt-contrib-qunit/issues/34