Update the documentation for the Mouse component
It needed to be more clear that you have to include the Mouse component before you can bind to Mouse events, else you could expect it to be like jQuery and wonder why i do not trigger anything.
Removed the text "the entity" from the triggers because it makes it more easier to read and you do understand that it talks about the entity.
Included a better link for the MouseEvent details.
Updated with a working example you can copy and paste.