Fix particles when viewport scaled
In the current version particles do not consider the current viewport zoom
when drawing themselve to the canvas.
This PR fixes that and also adds an option to specify an offset from the
base entity it is attached to.
original contribution in pr #743 by @johnHacksworth .
additional documentation & misc added by me.