Zoom scale fix
This fix changes zoom functionality in crafty

The current zoom function uses multiplication when computing the final
zoom amount.  This is fine except when computing zoom scales that are
not whole numbers.

This fix changes that to, instead of multiplication, use simple addition
and subtraction.  Therefore, it is now possible to use fractional scales
(0.1 - 0.9).

I have not fully tested the change across all of crafty but it does seem
to fix a bug in my particular code.