Refactor "Delay" to use dt from frameData
Rather than explicitly requesting the current time, the "Delay" component should just accumulate the time passed in by the "EnterFrame" event.  This way it doesn't need to worry about tracking the pause state of the game; that's all handled by the main game loop.

This meant drastically simplifying the tests -- they no longer need to overwrite the Date object or anything like that.