Make mobile act more like standard mode
Currently we don't handle the offset and positioning of cr-stage on mobile the same way we do in regular/desktop mode.  As far as I can see there's no actual reason for this.  (Perhaps it's a legacy of when mobile was forced into fullscreen mode?)

This patch handles the offset/position regardless of whether we're on mobile or not.  This fixes some graphical issues with Chrome/Firefox on Android, but I don't have an iOS device handy to test.  It should hopefully also fix an issue with touche coordinates being registered incorrectly.