Remove space character that affected layout
There was a non-breaking space character after the jquery script tag.

This was actually affecting the layout of the page, which I noticed when attempting to trigger mouse events.  Ironically, probably one of the more frustrating bugs to diagnose related Crafty to date.  As you might imagine, "is there an invisible character in the source code" was not high on my list of things to check when layout was being weird.  :)