Collect dom rendering system in one object, Crafty.domLayer in dom-layer.js; create domHelper
Previously some of the logic for creating the dom layer was in odd places; some of it was in the viewport code, some of it was in Crafty.DOM, while the actual rendering code was in Crafty.DrawManager.  This moves both the init and render functions into the same object, and places that in its own file.

It also takes the confusingly named Crafty.DOM and renames it Crafty.domHelper.