Remove test for Crafty.stop()
Unfortunately, running Crafty.stop() destroys a few types of bound events that aren't set back up on Crafty.init().

Long term, that should be fixed, but for now it shouldn't stop other tests from passing.

There was a work around for this issue in `tests/controls.js` that this removes.