Reduce array creation in Crafty.addComponent
`addComponent` created an array for iterating through the listed components.  But in each of the three possible cases, this was a duplication of effort; whether we're splitting the string or referring to the arguments, we already *have* an array like object to work with.

So with this patch, if the components are passed as arguments individually, we just use the arguments array; otherwise, we use the results of String.split.