Abstract out common event operations
There was a lot of duplicated logic between the various event-related functions.

This patch tries to abstract out the operations they had in common.  The basic idea is that a lot of the logic for bind, unbind, and trigger had to do with iterating through the callback array.   So this defines a set of operations for doing so, and then implements bind/unbind/trigger in terms of those operations.

Other features of this patch
- Give the Crafty object an "id"; `Crafty[0] == "global"` now.  Previously this had to be special cased in the event logic.
- Create a function `addCallbackMethods` which extends an object with the new low level callback methods.  The object itself still needs to provide bind/unbind/trigger/etc
- Iterate through collections of objects, rather than calling `.each()` with a new anonymous function.

In addition to cleaning up some of the existing code, this will make it much easier to create new objects which can listen to events.  I have a partial idea for how to make that even eaiser, but it'll need some work!