Introduce Crafty.systems
It's very natural to want an arbitrary, non-entity object that can participate in the event system.  Currently when objects like the viewport or the graphical layers want to listen to an event, they have to refer to themselves via hardcoded references, since the events happen in the global context.

This introduces the idea of systems which have a set of methods for participating in events.  These methods parallel those that Crafty and each entity implement.

When a system is created, a new object is created with these methods, and then it is extended with the methods/properties of a provided template object.  An `init` method and an `events` property can be provided that fill the same role as for components.  By default the system will be lazily created -- that is, it will only be created when referred to for the first time.  (This is what we do now with Crafty.canvasLayer and the like.)

In the future, we'll probably want to extend the capability of systems.  In particular, I envision that components and systems with the same name would have some connection -- perhaps the system could have methods for operating on all such entities at once.

I tested moving canvasLayer, domLayer, and webgl to this new paradigm and it seemed to work ok, but the actual switchover will happen in a new PR.