Change units in motion system
The motion system uses non-user friendly conversion between pixels
and the metric system. This commit changes all values to be in
pixels/frame and degrees/frame so that the user can easily access
and modify motion values without the need of external or internal
conversion between the units of measurement.
This commit also introduces the "FPSChange" event.