Fix some formatting issues in our doc blocks
We had inconsistent conventions regarding @param and @returns tags

- In some places, the description of the param/return value was continued across multiple lines
- In many other places, the description of the method as a whole was place directly after such tags

While the latter outnumbered the former, I think its useful to be able to have longer param/return descriptions split across multiple lines.  So to allow this while not breaking the docs, this fixes all incidences of the second style.  (This is probably also more readable when looking directly at the source.)