Sync motion with clock rather than number of frames
Instead of adjusting the speed based on the frame rate, try to saty in sync with the game clock by using frame.dt.  Note that speeds are in pixels per second.

This is a breaking change, because it changes how we interpret speeds.  (Pixels per second instead of pixels per frame.)