Fixed docs generation and added more examples.
There was a whitespace issue regarding documentation generation. Once that was adjusted everything is working fine.
I've also added a few examples to Crafty.extend and Crafty.clone. Extend works differently than most libraries, so it is good to have a simple example. And clone just didn't have one, so I made a very basic one for that.