Prevent gravity tunneling
Currently a falling object may fall through the floor if the gravity
constant is big or if the deltaTime between frames is large.
This commit fixes this by searching for collisions in the whole area a
falling entity might have skipped between frames.
This implementation is as (in)precise as the current implementation in
regards to choosing the ground entity among multiple ground candidates
that are close by.