Introduce UpdateFrame event
This event replaces  the EnterFrame event.
The EnterFrame event now triggers before the UpdateFrame event.
Replaces all triggered callbacks of the EnterFrame event to be triggered
by UpdateFrame instead.
Replaces the "EnterFrameInput" event triggered for the control system by
the "EnterFrame" event, as this now serves to process user input.