Change to the fourway component
I have tried to limit diagonal movement for the fourway component. Previously, the fourway component was a thin wrapper for multiway, which simpy adds each keypress to the vx and vy. With the changes I propose, the fourway component limits diagonal movement and does not add velocity but sets a new velocity with each key press, and check for keys still pressed at the keyup event setting the new velocity accordingly. I am using this implementation for an isometric game. If one were to use it for an rpg it would probably have to be limited further. Perhaps an fourway rpg component should be created as well.  

As a final note, I am unsure whether I handled the new binding of keydown and keyup in the simplest way. Let me know.