Handle parallax plus scaling more gracefully
When the viewport is scaled *and* layers have special x/y response values,
it's somewhat complicated to handle correctly.

This replaces the previous broken logic with a "correct" approach.
Under the new transform, a zoom in on the center of the viewport correctly
keeps layers from moving relative to each other. This unqiuely constrains
the properties under other camera movements.