Change TouchSystem
Change TouchSystem to inherit functionality from TouchState.
**Breaking change:**
* TouchStart and TouchEnd are no longer triggered
when a finger enters or leaves entity. They are now triggered only
once, when a finger is pressed or raised on the touch surface, no
matter which entity was targeted.
* TouchOver and TouchOut replace those events, which are triggered
when a finger enters or leaves the entity.

Add alias for Crafty.touchHandler.fingers.

Fix lifecycle onBlur events.