Fix for debugging non-renderable components
The debug layer assumed that any rendered components would have their own draw layer.
However, it's completely possible to have e.g. Collision components that aren't rendered,
and we want to be able to visualize these hitboxes as well.

When an entity doesn't have a specific draw layer, render it with the default viewport transforms.