Fix a bug with attached entities and origin change
When the origin of a rotated entity changes, we need to update the
position of any attached entities.

Rotation around an origin is equivalent to shifting to the origin,
rotating, and then shifting back.  When the origin changes, conceptually
we can handle this by reversing the current rotation, changing the origin,
and then rotating back.

However, this turns out to be equivavent to a simple shift in position,
dependent on the current rotation, and the shift in origin.
So th implementation calculates this shift and propagates it to child entities.